requirements.txt
django
djangorestframework
django-filter
requests